animate alpha value of background on detailed scan

+++ app/src/main/java/de/seemoo/at_tracking_detection/ui/scan/ScanDistanceViewModel.kt
@@ -8,6 +8,6 @@ class ScanDistanceViewModel : ViewModel() {
     var deviceAddress = MutableLiveData<String>()
     var connectionState = MutableLiveData<String>()
     var batteryState = MutableLiveData<String>()
-    var connectionQuality = MutableLiveData<Double>()
+    var connectionQuality = MutableLiveData<Int>()
 
 }
